Rudimentary support for updates of replicated arrays
Writes are just applied to each replica individually (in order). This
BREAKS sequential consistency.

Not sure, though, what consistency model it is. Without the use of
further synchronization primitves, it probably does not guarantee
anything. If barriers are used (such as with parallel OpenMP loops
without nowait), it probably implements release consistency.